requirements.txt
python-docx>=0.8.11
spacy>=3.7.2
nltk>=3.8.1
sentence-transformers>=2.2.2
torch>=2.0.0